Changed some names and reordered code
All math and other such operations on lists that were in the math class are now in the lists class. The math class has been renamed pyMath to avoid collision with python3s math library. Some functions have been moved like spline1D (was also renamed to perlin1D and overwrote that function sense they'er the same) and others. I also reordered the position of classes and functions to make more sense and labeled the chunks of code better.